requirements.txt
python-docx
pandas
docxcompose
docx2txt
deep-translator